refactor: unify utility modules
