+++ app/utils/responses.py
@@ -14,3 +14,11 @@ def error(message, status=400, code=None):
         "message": message,
         "code": code or status
     }), status
+
+
+def error_response(message: str, status_code: int):
+    return jsonify({"status": "error", "message": message}), status_code
+
+
+def internal_error_response():
+    return error_response("An unexpected error occurred, please try again later", 500)
